% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{athlet}
\alias{athlet}
\title{variable                         label
1    dscore  home scr. - vist. scr., 1993
2    dinstt    diff. in-state tuit., 1994
3   doutstt   diff. out-state tuit., 1994
4    htpriv    =1 if home team priv. sch.
5    vtpriv   =1 if vist. team priv. sch.
6     dapps   diff. in applications, 1994
7     htwrd  =1 if home win. record, 1993
8     vtwrd =1 if vist. win. record, 1993
9   dwinrec                 htwrd - vtwrd
10    dpriv               htpriv - vtpriv}
\format{
An object of class \code{data.frame} with 30 rows and 10 columns.
}
\usage{
athlet
}
\description{
variable                         label
1    dscore  home scr. - vist. scr., 1993
2    dinstt    diff. in-state tuit., 1994
3   doutstt   diff. out-state tuit., 1994
4    htpriv    =1 if home team priv. sch.
5    vtpriv   =1 if vist. team priv. sch.
6     dapps   diff. in applications, 1994
7     htwrd  =1 if home win. record, 1993
8     vtwrd =1 if vist. win. record, 1993
9   dwinrec                 htwrd - vtwrd
10    dpriv               htpriv - vtpriv
}
\keyword{datasets}
